third_party/WebKit/LayoutTests/bluetooth/requestDevice/name-empty-device-from-name-empty-filter.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setUpPreconnectedDevice({name: ''})
    .then(() => requestDeviceWithTrustedClick({filters: [{name: ''}]}))
    .then(device => {
      assert_equals(device.name, '');
    });
}, 'An empty name device can be obtained by empty name filter.');
</script>
